# [4.x] Improve collection Stache watcher performance