--- a/src/Stache/Stores/CollectionsStore.php
+++ b/src/Stache/Stores/CollectionsStore.php
@@ -3,6 +3,7 @@
 namespace Statamic\Stache\Stores;
 
 use Statamic\Facades\Collection;
+use Statamic\Facades\Path;
 use Statamic\Facades\Site;
 use Statamic\Facades\Stache;
 use Statamic\Facades\YAML;
@@ -24,7 +25,10 @@ class CollectionsStore extends BasicStore
 
     public function getItemFilter(SplFileInfo $file)
     {
-        return $file->getExtension() === 'yaml';
+        $dir = str_finish($this->directory, '/');
+        $relative = str_after(Path::tidy($file->getPathname()), $dir);
+
+        return $file->getExtension() === 'yaml' && substr_count($relative, '/') === 0;
     }
 
     public function makeItemFromFile($path, $contents)