--- a/src/Stache/Stores/Store.php
+++ b/src/Stache/Stores/Store.php
@@ -22,6 +22,7 @@ abstract class Store
     protected $storeIndexes = [];
     protected $usedIndexes;
     protected $fileChangesHandled = false;
+    protected $fileChangesAreRecursive = true;
     protected $paths;
     protected $fileItems;
     protected $shouldCacheFileItems = false;
@@ -180,7 +181,7 @@ abstract class Store
         $existing = collect(Cache::get($cacheKey, []));
 
         // Get the files and timestamps from the filesystem right now.
-        $files = Traverser::filter([$this, 'getItemFilter'])->traverse($this);
+        $files = Traverser::filter([$this, 'getItemFilter'])->traverse($this, $this->fileChangesAreRecursive);
 
         // Cache the files and timestamps, ready for comparisons on the next request.
         // We'll do it now since there are multiple early returns coming up.