--- a/src/Stache/Stores/Store.php
+++ b/src/Stache/Stores/Store.php
@@ -22,7 +22,7 @@ abstract class Store
     protected $storeIndexes = [];
     protected $usedIndexes;
     protected $fileChangesHandled = false;
-    protected $fileChangesAreRecursive = true;
+    protected $traverseRecursively = true;
     protected $paths;
     protected $fileItems;
     protected $shouldCacheFileItems = false;
@@ -181,7 +181,7 @@ abstract class Store
         $existing = collect(Cache::get($cacheKey, []));
 
         // Get the files and timestamps from the filesystem right now.
-        $files = Traverser::filter([$this, 'getItemFilter'])->traverse($this, $this->fileChangesAreRecursive);
+        $files = Traverser::filter([$this, 'getItemFilter'])->traverse($this, $this->traverseRecursively);
 
         // Cache the files and timestamps, ready for comparisons on the next request.
         // We'll do it now since there are multiple early returns coming up.
@@ -293,7 +293,7 @@ abstract class Store
             return $this->paths = collect($paths);
         }
 
-        $files = Traverser::filter([$this, 'getItemFilter'])->traverse($this, $this->fileChangesAreRecursive);
+        $files = Traverser::filter([$this, 'getItemFilter'])->traverse($this, $this->traverseRecursively);
 
         $fileItems = $files->map(function ($timestamp, $path) {
             return [