This PR optimizes the Stache watcher by limiting the scope of the Collection Stache watcher to just the immediate files within the `content/collections` directory. The changes allow for any store to disable recursive file retrieval (currently only enabled on the collection store).

From testing, the existing behavior will load *all* file paths within the `content/` directory when checking for Collection changes, and then load them all again for each individual collection that is checked.

The following #'s are not super rigorous, and are just general observations from viewing a single article page in the browser (site based on Cool Writings). All numbers are after warming the Stache separately before running the site.

Baseline numbers without any extra collection entries: bounce between 750ms to 1.3s

| Additional Entries | Before TTFB | After TTFB |
|---|---|---|
| 15,109 | ~4.1s | ~1.9s |

These numbers remain fairly consistent event when increasing the number of file based entries to 20K+. Site is running on Windows via. WSL without any environment optimizations and using `php artisan serve`.

**Note 1**: While working through this, I had the thought that it would be possible to disable the Stache watcher for certain collections, and can look into that if there is interest (i.e., disable it on very large collections with infrequent updates, but keep it enabled for smaller ones, etc.)

**Note 2**: The back/forth changes related to removing the string operations inside the collection store were to preserve any existing behavior that relies purely on the store's `getFilterItem` method. Removing these would improve things very, *very* slightly, but would definitely break behavior at this time (hooray existing test coverage).