@jasonvarga 2024-01-30
I'm also testing on the Cool Writings /articles listing. It paginates to the first 10 items.
I have 20k entries. The page with the Stache watcher enabled loads at about ~1.7s. With this PR it's about ~1.2s.

With the Stache watcher disabled there's no difference - which is expected.

I think the drastic difference is due to you using Windows. (I'm sorry!)

Still, small code change results in a not-nothing performance increase. 👍