@aerni 2024-02-01
This PR broke the custom stache store in my [Advanced SEO](https://github.com/aerni/statamic-advanced-seo) addon. I get unlimited duplicate ID's. Regenerating just adds another duplicate. Do I have to change something on my side or is this a bug or breaking change?

<img width="946" alt="CleanShot 2024-02-01 at 12 11 22@2x" src="https://github.com/statamic/cms/assets/23167701/0153ce14-65da-41fd-b058-4ab1f8676ce5">