@jasonvarga 2024-02-01
I'll revert it. Shouldn't have been a breaking change. Sorry about that.